requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
pyyaml
typer